+++ CHANGES.md
@@ -1,5 +1,9 @@
 # Change Log
 
+*   2.1.0
+    *   Add [Super CSV](http://super-csv.github.io/super-csv/)-based `TabularDataSource`
+    *   Deprecate opencsv-based `TabularDataSource`; the Super CSV-based `TabularDataSource` should be used instead.
+
 *   2.0.1
     *   Upgrade to POI 3.10.1